Themis version: master 0.0.2 0.0.1

Struct themis::secure_session::SecureSession

pub struct SecureSession<T> { /* fields omitted */ }

Secure Session context.

Methods

impl<T> SecureSession<T> where
T: SecureSessionTransport,

pub fn with_transport<I>(
id: I,
key: &EcdsaSecretKey,
transport: T
) -> Result<Self> where
I: AsRef<[u8]>,

Creates a new Secure Session.

ID is an arbitrary byte sequence used to identify this peer.

Secure Session supports only ECDSA keys.

pub fn is_established(&self) -> bool

Returns true if this Secure Session may be used for data transfer.

pub fn get_remote_id(&self) -> Result<Vec<u8>>

Returns ID of the remote peer.

This method will return an error if the connection has not been established yet.

pub fn connect(&mut self) -> Result<()>

Initiates connection to the remote peer.

This is the first method to call. It uses transport callbacks to send the resulting connection request to the peer. Afterwards call negotiate_transport until the connection is established. That is, until the state_changed callback of your SecureSessionTransport tells you that the connection is Established, or until is_established on this Secure Session returns true.

This method is a part of transport API and requires send_data method of SecureSessionTransport.

pub fn generate_connect_request(&mut self) -> Result<Vec<u8>>

Initiates connection to the remote peer, returns connection message.

This is the first method to call. It returns you a message that you must somehow transfer to the remote peer and give it to its negotiate method. This results in another message which must be transferred back to this Secure Session and passed to its negotiate method. Continue passing these message around until the connection is established. That is, until the state_changed callback of your SecureSessionTransport tells you that the connection is Established, or until is_established on this Secure Session returns true, or until negotiate returns an empty message.

pub fn wrap<M: AsRef<[u8]>>(&mut self, message: M) -> Result<Vec<u8>>

Wraps a message and returns it.

The message can be transferred to the remote peer and unwrapped there with unwrap.

Wrapped message are independent and can be exchanged out of order. You can wrap multiple messages then unwrap them in any order or don't unwrap some of them at all.

This method will fail if a secure connection has not been established yet.

pub fn unwrap<M: AsRef<[u8]>>(&mut self, wrapped: M) -> Result<Vec<u8>>

Unwraps a message and returns it.

Unwraps a message previously wrapped by the remote peer.

This method will fail if a secure connection has not been established yet.

pub fn negotiate<M: AsRef<[u8]>>(&mut self, wrapped: M) -> Result<Vec<u8>>

Continues connection negotiation with given message.

This method performs one step of connection negotiation. The server should call this method first with a message received from client's generate_connect_request. Its result is another negotiation message that should be transferred to the client. The client then calls this method on a message and forwards the resulting message to the server. If the returned message is empty then negotiation is complete and the Secure Session is ready to be used.

pub fn send<M: AsRef<[u8]>>(&mut self, message: M) -> Result<()>

Sends a message to the remote peer.

This method will fail if a secure connection has not been established yet.

This method is a part of transport API and requires send_data method of SecureSessionTransport.

pub fn receive(&mut self, max_len: usize) -> Result<Vec<u8>>

Receives a message from the remote peer.

Maximum length of the message is specified by the parameter. The message will be truncated to this length if the peer sends something larger.

This method will fail if a secure connection has not been established yet.

This method is a part of transport API and requires receive_data method of SecureSessionTransport.

pub fn negotiate_transport(&mut self) -> Result<()>

Continues connection negotiation.

This method performs one step of connection negotiation. This is the first method to call for the server, the client calls it after the initial connect. Both peers shall then repeatedly call this method until the connection is established (see connect for details).

This method is a part of transport API and requires send_data and receive_data methods of SecureSessionTransport.