Themis version: master 0.0.2 0.0.1

Trait themis::secure_session::SecureSessionTransport

pub trait SecureSessionTransport {
    fn get_public_key_for_id(&mut self, id: &[u8], key: &mut [u8]) -> bool;

    fn send_data(&mut self, data: &[u8]) -> Result<usize, ()> { ... }

    fn receive_data(&mut self, data: &mut [u8]) -> Result<usize, ()> { ... }

    fn state_changed(&mut self, state: SecureSessionState) { ... }
}

Transport delegate for Secure Session.

This is an interface you need to provide for Secure Session operation.

The only required method is get_public_key_for_id. It is required for public key authentication. Other methods are optional, you can use Secure Session without them, but some functionality may be unavailable.

Required Methods

fn get_public_key_for_id(&mut self, id: &[u8], key: &mut [u8]) -> bool

Get a public key corresponding to a peer ID.

Look up the key and place it into the provided buffer. Return true if a key has been found for the provided ID and false for unknown IDs.

Provided Methods

fn send_data(&mut self, data: &[u8]) -> Result<usize, ()>

Send the provided data to the peer, return the number of bytes transferred.

This callback will be called when Secure Session needs to send some data to its peer. The whole message is expected to be transferred so returning anything other than Ok(data.len()) is considered an error.

This method is used by the transport API (connect, negotiate_transport, send). You need to implement it in order to use this API.

fn receive_data(&mut self, data: &mut [u8]) -> Result<usize, ()>

Receive some data from the peer into the provided buffer, return the number of bytes.

This callback will be called when Secure Session expects to receive some data. The length of the buffer indicates the maximum amount of data expected. Put the received data into the provided buffer and return the number of bytes that you used.

This method is used by the transport API (negotiate_transport, receive). You need to implement it in order to use this API.

fn state_changed(&mut self, state: SecureSessionState)

Notification about connection state of Secure Session.

This method is truly optional and has no effect on Secure Session operation.